pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wikirss"
version = "0.1.0"
description = "A scraper that converts the Wikipedia front page to a daily RSS feed."
authors = [
    { name = "Henry Stern"},
]
license = { file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedgen>=1.0.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = [
    "W",                   # Whitespace
    "N",                   # Naming
    "A",                   # Overshadowing builtins
    "ISC",                 # Implicit string concatenation
    "PTH",                 # Require pathlib instead of os
    "D"                    # Require docstrings
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = ["D104"] # Don't require docstring in __init__

[tool.ruff.lint.isort]
known-first-party = ["wikirss"]
force-sort-within-sections = true

[tool.pyright]
typeCheckingMode = "standard"
include = ["src"]
ignore = ["\\Interactive-*.interactive.py"]
reportMissingParameterType = true

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "nbformat>=5.10.4",
]
